Allow a 'class' instance in the listener instead of inline definitions
I'm using TypeScript and I wanted to implement the client as a class (and its interface) in a separatedly place where Hub definition is.
Setting up the listener property with 'new Client(args)' instead of defining the object inline, It didn't work. I think because angular.foreach skip prototypes properties. So, with this change i think it does it ok.